Port TestRenameNodeWiki
